requirements.txt
flask
python-dotenv
flask-cors
flask-mysql
pyjwt